refactor: standardization adjustments
