+++ src/app/libs/axios/api.ts
@@ -1,7 +1,5 @@
 import axios from 'axios'
 
-const api = axios.create({
+export const api = axios.create({
   baseURL: 'http://localhost:3333',
 })
-
-export default api
