+++ src/app/libs/tanstack-query/index.ts
@@ -1,1 +1,2 @@
+export * from './query-client'
 export * from './tanstack-query-provider'
