+++ src/app/libs/tanstack-query/query-client.ts
@@ -1,0 +1,3 @@
+import { QueryClient } from '@tanstack/react-query'
+
+export const queryClient = new QueryClient()
