+++ src/app/libs/tanstack-query/tanstack-query-provider.tsx
@@ -1,11 +1,16 @@
 'use client'
 
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import { QueryClientProvider } from '@tanstack/react-query'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
+import { queryClient } from '.'
 
-const TanstackQueryProvider = ({ children }: { children: React.ReactNode }) => {
-  const queryClient = new QueryClient()
+interface TanstackQueryProviderProps {
+  children: React.ReactNode
+}
 
+export const TanstackQueryProvider = ({
+  children,
+}: TanstackQueryProviderProps) => {
   return (
     <QueryClientProvider client={queryClient}>
       {children}
@@ -13,5 +18,3 @@ const TanstackQueryProvider = ({ children }: { children: React.ReactNode }) => {
     </QueryClientProvider>
   )
 }
-
-export default TanstackQueryProvider
